moved the python code to the src directory
